+++ tests/test_commands.py
@@ -15,6 +15,7 @@ class SeedingCommandTestCase(BaseCommandTestCase):
     def test_seeding_random(self):
         # Prerequisites
         self.runner.invoke(args=['import', 'categories', 'tests/resources/science-categories.json'])
+        self.runner.invoke(args=['import', 'organisations', 'tests/resources/funder-organisations.json'])
 
         result = self.runner.invoke(args=['seed', 'random'])
         self.assertIn('Seeded random mock resources', result.output)
@@ -24,3 +25,8 @@ class ImportCommandTestCase(BaseCommandTestCase):
     def test_import_categories_from_file(self):
         result = self.runner.invoke(args=['import', 'categories', 'tests/resources/science-categories.json'])
         self.assertIn('Finished importing research categories', result.output)
+
+    def test_import_organisations_from_file(self):
+        result = self.runner.invoke(args=['import', 'organisations', 'tests/resources/funder-organisations.json'])
+        self.assertIn('Finished importing organisations', result.output)
+